Refactored DataGrid component.
