+++ samples/.eslintrc
@@ -6,6 +6,7 @@ globals:
     ADBE: true
     CallAnalyzer: true
     Dashboard: true
+    DataGrid: true
     GOOGL: true
     Highcharts: true
     hs: true                    # Highslide
