+++ samples/highcharts/studies/dashboard-datagrid-component/demo.html
@@ -3,6 +3,7 @@
 <script src="https://code.highcharts.com/dashboard.js"></script>
 <script src="https://code.highcharts.com/highcharts.js"></script>
 <script src="https://code.highcharts.com/datagrid.js"></script>
+<script src="https://code.highcharts.com/modules/draggable-points.js"></script>
 
 <div id="container" class="highcharts-dashboard">
 <div id="datagrid" ></div>
