+++ samples/tsglobal.d.ts
@@ -1,4 +1,6 @@
 /// <reference path="../code/highcharts.d.ts" />
+/// <reference path="../code/dashboard.d.ts" />
+/// <reference path="../code/datagrid.d.ts" />
 /// <reference path="../node_modules/@types/jquery/index.d.ts" />
 /// <reference path="../node_modules/@types/qunit/index.d.ts" />
 /// <reference path="../test/test-controller.d.ts" />
