+++ test/typescript-karma/Extensions/DashboardPlugin/tests.js
@@ -1,18 +1,17 @@
 import CSVStore from '/base/code/es-modules/Data/Stores/CSVStore.js';
 import PluginHandler from '/base/code/es-modules/Dashboard/PluginHandler.js';
 import Highcharts from '/base/code/es-modules/masters/highcharts.src.js';
-import HighchartsComponent from '/base/code/es-modules/Extensions/DashboardPlugin/HighchartsComponent.js';
-import HighchartsPlugin from '/base/code/es-modules/Extensions/DashboardPlugin/HighchartsPlugin.js';
-import DataGrid from '/base/js/DataGrid/DataGrid.js';
-import DataGridComponent from '/base/code/es-modules/Extensions/DashboardPlugin/DataGridComponent.js';
-import DataGridPlugin from '/base/code/es-modules/Extensions/DashboardPlugin/DataGridPlugin.js';
+import HighchartsComponent from '/base/code/es-modules/Extensions/DashboardPlugins/HighchartsComponent.js';
+import HighchartsPlugin from '/base/code/es-modules/Extensions/DashboardPlugins/HighchartsPlugin.js';
+import DataGrid from '/base/code/es-modules/DataGrid/DataGrid.js';
+import DataGridComponent from '/base/code/es-modules/Extensions/DashboardPlugins/DataGridComponent.js';
+import DataGridPlugin from '/base/code/es-modules/Extensions/DashboardPlugins/DataGridPlugin.js';
 
 HighchartsPlugin.custom.connectHighcharts(Highcharts);
 PluginHandler.addPlugin(HighchartsPlugin);
 
 DataGridPlugin.custom.connectDataGrid(DataGrid);
 PluginHandler.addPlugin(DataGridPlugin);
-
 const { test } = QUnit;
 
 test('Both components should work', (assert) => {
@@ -113,8 +112,7 @@ test('HighchartsComponent constructors', function (assert) {
 });
 
 test('DataGridComponent constructors', function (assert) {
-    console.log(DataGridComponent);
-    const component = new DataGridComponent({});
+    const component = new DataGridComponent({}).render();
     // Test that the constructor creates a component.
-    assert.ok(component, `Able to create a DataGridComponent.`);
+    assert.ok(component.dataGrid, `Able to create a DataGridComponent.`);
 });
