+++ ts/Extensions/DashboardPlugins/HighchartsComponent.ts
