+++ ts/Extensions/DashboardPlugins/HighchartsPlugin.ts
