+++ ts/Extensions/DashboardPlugins/HighchartsSyncHandlers.ts
